+++ source/Statistics.cpp
@@ -1268,7 +1268,7 @@ void Statistics::printPhaseResultsLatencyToStream(const LatencyHistogram& latHis
 		// individual results header (note: keep format in sync with general table format string)
 		outStream << boost::format(Statistics::phaseResultsLeftFormatStr)
 			% ""
-			% (latTypeStr + " lat %ile")
+			% (latTypeStr + " lat %ile us")
 			% ":";
 
 		outStream << "[ ";
@@ -1276,11 +1276,26 @@ void Statistics::printPhaseResultsLatencyToStream(const LatencyHistogram& latHis
 		if(latHisto.getHistogramExceeded() )
 			outStream << "Histogram exceeded";
 		else
+		{
 			outStream <<
 				"1%<=" << latHisto.getPercentileStr(1) << " "
 				"50%<=" << latHisto.getPercentileStr(50) << " "
 				"75%<=" << latHisto.getPercentileStr(75) << " "
 				"99%<=" << latHisto.getPercentileStr(99);
+
+			// print configurable amount of percentile nines (99.9%, 99.99%, ...)
+			std::string ninesStr = "99.";
+			for(unsigned short numDecimals=1;
+				numDecimals <= progArgs.getNumLatencyPercentile9s();
+				numDecimals++)
+			{
+				ninesStr += "9"; // append next decimal 9
+				double percentage = std::stod(ninesStr);
+
+				outStream << " " << std::setprecision(numDecimals+3) << // +3 for "99."
+					percentage << "%<=" << latHisto.getPercentileStr(percentage);
+			}
+		}
 
 		outStream << " ]" << std::endl;
 	}